+++ qsprpred/data/utils/descriptorsets.py
@@ -1,4 +1,4 @@
-"""Descriptorssets.
+"""Descriptorssets. A descriptorset is a collection of descriptors that can be calculated for a molecule.
 
 To add a new descriptor or fingerprint calculator:
 * Add a descriptor subclass for your descriptor calculator
@@ -6,12 +6,13 @@ To add a new descriptor or fingerprint calculator:
 """
 
 from abc import ABC, abstractmethod
-from typing import Union, List
+from typing import List, Union
 
 import mordred
 import numpy as np
 import pandas as pd
 from mordred import descriptors as mordreddescriptors
+from qsprpred.data.utils.descriptor_utils import fingerprints
 from qsprpred.data.utils.descriptor_utils.drugexproperties import Property
 from qsprpred.data.utils.descriptor_utils.rdkitdescriptors import RDKit_desc
 from rdkit import Chem, DataStructs
@@ -19,10 +20,13 @@ from rdkit.Chem import AllChem, Mol
 
 
 class DescriptorSet(ABC):
-    """Abstract base class for descriptorssets."""
+    """Abstract base class for descriptorsets.
+
+    A descriptorset is a collection of descriptors that can be calculated for a molecule.
+    """
 
     @abstractmethod
-    def __call__(self, mols : List[Union[str, Mol]]):
+    def __call__(self, mols: List[Union[str, Mol]]):
         """
         Calculate the descriptor for a molecule.
 
@@ -34,7 +38,7 @@ class DescriptorSet(ABC):
         """
         pass
 
-    def iterMols(self, mols : List[Union[str, Mol]], to_list=False):
+    def iterMols(self, mols: List[Union[str, Mol]], to_list=False):
         """
         Calculate the descriptor for a molecule.
 
@@ -44,7 +48,6 @@ class DescriptorSet(ABC):
         Returns:
             an array or data frame of descriptor values of shape (n_mols, n_descriptors)
         """
-
         ret = (Chem.MolFromSmiles(mol) if isinstance(mol, str) else mol for mol in mols)
         if to_list:
             ret = list(ret)
@@ -63,6 +66,7 @@ class DescriptorSet(ABC):
         pass
 
     def get_len(self):
+        """Return the number of descriptors."""
         return len(self.descriptors)
 
     @abstractmethod
@@ -72,37 +76,40 @@ class DescriptorSet(ABC):
 
     @abstractmethod
     def settings(self):
-        """Return args and kwargs used to initialize the descriptorset."""
+        """Return dictionary with arguments used to initialize the descriptorset."""
         pass
 
     @abstractmethod
     def __str__(self):
+        """Return string representation of the descriptorset."""
         pass
 
 
-class MorganFP(DescriptorSet):
+class FingerprintSet(DescriptorSet):
+    """Generic fingerprint descriptorset can be used to calculate any fingerprint type defined in descriptorutils.fingerprints."""
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, fingerprint_type, *args, **kwargs):
         """
-        Initialize the descriptor with the same arguments as you would pass to `GetMorganFingerprintAsBitVect` function of RDKit.
+        Initialize the descriptor with the same arguments as you would pass to your fingerprint type of choice.
 
         Args:
-            *args: `GetMorganFingerprintAsBitVect` arguments
-            **kwargs: `GetMorganFingerprintAsBitVect` keyword arguments
+            fingerprint_type: fingerprint type
+            *args: fingerprint specific arguments
+            **kwargs: fingerprint specific arguments keyword arguments
         """
-        self._args = args
-        self._kwargs = kwargs
         self._is_fp = True
+        self.fingerprint_type = fingerprint_type
+        self.get_fingerprint = fingerprints.get_fingerprint(self.fingerprint_type, *args, **kwargs)
 
         self._keepindices = None
 
     def __call__(self, mols):
+        """Calculate the fingerprint for a list of molecules."""
         convertFP = DataStructs.ConvertToNumpyArray
-        morgan = AllChem.GetMorganFingerprintAsBitVect
 
         ret = np.zeros((len(mols), self.get_len()))
-        for idx,mol in enumerate(self.iterMols(mols)):
-            fp = morgan(mol, *self._args, **self._kwargs)
+        for idx, mol in enumerate(self.iterMols(mols)):
+            fp = self.get_fingerprint(mol)
             np_fp = np.zeros(len(fp))
             convertFP(fp, np_fp)
             if self.keepindices:
@@ -113,60 +120,86 @@ class MorganFP(DescriptorSet):
 
     @property
     def keepindices(self):
+        """Return the indices of the fingerprint to keep."""
         return self._keepindices
 
     @keepindices.setter
     def keepindices(self, val):
+        """Set the indices of the fingerprint to keep."""
         self._keepindices = [int(x) for x in val] if val else None
 
     @property
     def is_fp(self):
+        """Return True if descriptorset is fingerprint."""
         return self._is_fp
 
     @property
     def settings(self):
-        return self._args, self._kwargs
+        """Return dictionary with arguments used to initialize the descriptorset."""
+        return {"fingerprint_type": self.fingerprint_type, **self.get_fingerprint.settings}
 
     def get_len(self):
-        return len(AllChem.GetMorganFingerprintAsBitVect(Chem.MolFromSmiles("C"), *self._args, **self._kwargs))
+        """Return the length of the fingerprint."""
+        return len(self.get_fingerprint(Chem.MolFromSmiles("C")))
 
     def __str__(self):
-        return "MorganFP"
+        return f"FingerprintSet"
 
     @property
     def descriptors(self):
+        """Return the indices of the fingerprint that are kept."""
         indices = self.keepindices if self.keepindices else range(self.get_len())
         return [f"{idx}" for idx in indices]
 
     @descriptors.setter
     def descriptors(self, value):
-        """
-        Sets the number of bits to a given value.
-        """
-        self._kwargs["nBits"] = value
+        """Set the indices of the fingerprint to keep."""
+        self.keepindices(value)
 
 
 class Mordred(DescriptorSet):
     """Descriptors from molecular descriptor calculation software Mordred.
 
     From https://github.com/mordred-descriptor/mordred.
-    Initialize the descriptor with the same arguments as you would pass to `Calculator` function of Mordred.
-    If no mordred descriptor object passed, the all descriptors will be calculated.
 
     Args:
-        *args: `Calculator` arguments
-        **kwargs: `Calculator` keyword arguments
+        descs (list): list of mordred descriptor names
+        version (str): version of mordred
+        ignore_3D (bool): ignore 3D information
+        config (str): path to config file
     """
 
-    def __init__(self, *args, **kwargs):
-        self._args = args
-        self._kwargs = kwargs
-        self._process_args(*args, **kwargs)
+    def __init__(self, descs=None, version=None, ignore_3D=False, config=None):
+        """
+        Initialize the descriptor with the same arguments as you would pass to `Calculator` function of Mordred.
+
+        With the exception of the `descs` argument which can also be a list of mordred descriptor names instead
+        of a mordred descriptor module.
+
+        Args:
+            descs: List of Mordred descriptor names, a Mordred descriptor module or None for all mordred descriptors
+            version (str): version of mordred
+            ignore_3D (bool): ignore 3D information
+            config (str): path to config file?
+        """
+        if descs:
+            # if mordred descriptor module is passed, convert to list of descriptor instances
+            if not isinstance(descs, list):
+                descs = (mordred.Calculator(descs).descriptors)
+        else:
+            # use all mordred descriptors if no descriptors are specified
+            descs = mordred.Calculator(mordreddescriptors).descriptors
+
+        self.version = version
+        self.ignore_3D = ignore_3D
+        self.config = config
 
         self._is_fp = False
 
         self._mordred = None
-        self.descriptors = self._args[0]
+
+        # convert to list of descriptor names if descriptor instances are passed and initiate mordred calulator
+        self.descriptors = [str(d) for d in descs]
 
     def __call__(self, mols):
         return self._mordred.pandas(self.iterMols(mols), quiet=True, nproc=1).values
@@ -177,7 +210,7 @@ class Mordred(DescriptorSet):
 
     @property
     def settings(self):
-        return self._args, self._kwargs
+        return {"descs": self.descriptors, "version": self.version, "ignore_3D": self.ignore_3D, "config": self.config}
 
     @property
     def descriptors(self):
@@ -185,20 +218,19 @@ class Mordred(DescriptorSet):
 
     @descriptors.setter
     def descriptors(self, names):
+        """Set the descriptors to calculate.
+
+        Converts a list of Mordred descriptor names to Mordred descriptor instances which is used to initialize the
+        a Mordred calculator with the specified descriptors.
+
+        Args:
+            names: List of Mordred descriptor names.
+        """
         calc = mordred.Calculator(mordreddescriptors)
         self._mordred = mordred.Calculator(
-            [d for d in calc.descriptors if str(d) in names], **self._kwargs
-        )
+            [d for d in calc.descriptors if str(d) in names],
+            version=self.version, ignore_3D=self.ignore_3D, config=self.config)
         self._descriptors = names
-
-    def _process_args(self, descs=None, version=None, ignore_3D=False, config=None):
-        if descs:
-            if not isinstance(descs, list):
-                descs = (mordred.Calculator(descs).descriptors)
-        else:
-            descs = mordred.Calculator(mordreddescriptors).descriptors
-        self._args = [[str(d) for d in descs]]
-        self._kwargs = {"version": version, "ignore_3D": ignore_3D, "config": config}
 
     def __str__(self):
         return "Mordred"
@@ -206,21 +238,23 @@ class Mordred(DescriptorSet):
 
 class DrugExPhyschem(DescriptorSet):
     """
-    Pysciochemical properties originally used in DrugEx for QSAR modelling
-    Initialize the descriptor with Property arguments (a list of properties to calculate) to select a subset.
+    Pysciochemical properties originally used in DrugEx for QSAR modelling.
 
     Args:
-        *args: `Property` arguments
-        **kwargs: `Property` keyword arguments
+        props: list of properties to calculate
     """
 
-    def __init__(self, *args, **kwargs):
-        self._args = args
-        self._kwargs = kwargs
+    def __init__(self, physchem_props=None):
+        """Initialize the descriptorset with Property arguments (a list of properties to calculate) to select a subset.
+
+        Args:
+            physchem_props: list of properties to calculate
+        """
         self._is_fp = False
-        self.props = [x for x in Property(*args, **kwargs).props]
+        self.props = [x for x in Property(physchem_props).props]
 
     def __call__(self, mols):
+        """Calculate the DrugEx properties for a molecule."""
         calculator = Property(self.props)
         return calculator.getScores(self.iterMols(mols, to_list=True))
 
@@ -230,7 +264,7 @@ class DrugExPhyschem(DescriptorSet):
 
     @property
     def settings(self):
-        return self._args, self._kwargs
+        return {"physchem_props": self.props}
 
     @property
     def descriptors(self):
@@ -247,21 +281,18 @@ class DrugExPhyschem(DescriptorSet):
 
 class rdkit_descs(DescriptorSet):
     """
-    RDkit descriptors
-    Initialize the descriptor names (a list of properties to calculate) to select a subset of the rdkit descriptors.
-    Add compute_3Drdkit argument to indicate if 3D descriptors should also be calculated
+    Calculate RDkit descriptors.
 
     Args:
-        *args: `Property` arguments
-        **kwargs: `Property` keyword arguments
+        rdkit_descriptors: list of descriptors to calculate, if none, all 2D rdkit descriptors will be calculated
+        compute_3Drdkit: if True, 3D descriptors will be calculated
     """
 
-    def __init__(self, *args, **kwargs):
-        self._args = args
-        self._kwargs = kwargs
+    def __init__(self, rdkit_descriptors=None, compute_3Drdkit=False):
         self._is_fp = False
-        self._calculator = RDKit_desc(*args, **kwargs)
+        self._calculator = RDKit_desc(rdkit_descriptors, compute_3Drdkit)
         self._descriptors = self._calculator.descriptors
+        self.compute_3Drdkit = compute_3Drdkit
 
     def __call__(self, mols):
         return self._calculator.getScores(self.iterMols(mols, to_list=True))
@@ -272,7 +303,7 @@ class rdkit_descs(DescriptorSet):
 
     @property
     def settings(self):
-        return self._args, self._kwargs
+        return {"rdkit_descriptors": self.descriptors, "compute_3Drdkit": self.compute_3Drdkit}
 
     @property
     def descriptors(self):
@@ -287,24 +318,95 @@ class rdkit_descs(DescriptorSet):
         return "RDkit"
 
 
+class TanimotoDistances(DescriptorSet):
+    """
+    Calculate Tanimoto distances to a list of SMILES sequences.
+
+    Args:
+        list_of_smiles (list of strings): list of SMILES sequences to calculate distance to
+        fingerprint_type (str): fingerprint type to use
+        *args: `fingerprint` arguments
+        **kwargs: `fingerprint` keyword arguments, should contain fingerprint_type
+    """
+
+    def __init__(self, list_of_smiles, fingerprint_type, *args, **kwargs):
+        """Initialize the descriptorset with a list of SMILES sequences and a fingerprint type.
+
+        Args:
+            list_of_smiles (list of strings): list of SMILES sequences to calculate distance to
+            fingerprint_type (str): fingerprint type to use
+        """
+        self._descriptors = list_of_smiles
+        self.fingerprint_type = fingerprint_type
+        self._args = args
+        self._kwargs = kwargs
+        self._is_fp = False
+
+        # intialize fingerprint calculator
+        self.get_fingerprint = fingerprints.get_fingerprint(self.fingerprint_type, *self._args, **self._kwargs)
+        self.fps = self.calculate_fingerprints(list_of_smiles)
+
+    def __call__(self, mols):
+        """Calculate the Tanimoto distances to the list of SMILES sequences.
+
+        Args:
+            mols (list of rdkit.Chem.rdchem.Mol): SMILES sequence or RDKit molecule to calculate distances to
+        """
+        ret = np.zeros((len(mols), self.get_len()))
+        for idx, mol in enumerate(self.iterMols(mols, to_list=True)):
+            ret[idx] = 1 - np.array(DataStructs.BulkTanimotoSimilarity(self.get_fingerprint(mol), self.fps))
+
+        return ret
+
+    def calculate_fingerprints(self, list_of_smiles):
+        """Calculate the fingerprints for the list of SMILES sequences."""
+        return [self.get_fingerprint(Chem.MolFromSmiles(smiles)) for smiles in list_of_smiles]
+
+    @property
+    def is_fp(self):
+        return self._is_fp
+
+    @property
+    def settings(self):
+        return {"fingerprint_type": self.fingerprint_type,
+                "list_of_smiles": self.descriptors, "args": self._args, "kwargs": self._kwargs}
+
+    @property
+    def descriptors(self):
+        return self._descriptors
+
+    @descriptors.setter
+    def descriptors(self, list_of_smiles):
+        """Set new list of SMILES sequences to calculate distance to."""
+        self._descriptors = list_of_smiles
+        self.fps = self.calculate_fingerprints(list_of_smiles)
+
+    def get_len(self):
+        return len(self.descriptors)
+
+    def __str__(self):
+        return "TanimotoDistances"
+
+
 class PredictorDesc(DescriptorSet):
     """DescriptorSet that uses a Predictor object to calculate the descriptors for a molecule."""
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, model_path: str, metadata_path: str):
         """
         Initialize the descriptorset with a Predictor object.
 
         Args:
-            predictor: Predictor object to use for calculating the descriptor
+            model_path: path to the model file
+            metadata_path: path to the metadata file
         """
-        self._args = args
-        self._kwargs = kwargs
         self._is_fp = False
+        self.model_path = model_path
+        self.metadata_path = metadata_path
         from qsprpred.scorers.predictor import Predictor
-        self._predictor = Predictor.fromFile(*args, **kwargs)
+        self._predictor = Predictor.fromFile(model_path, metadata_path)
         self._descriptors = [self._predictor.getKey()]
 
-    def __call__(self, mol):
+    def __call__(self, mols):
         """
         Calculate the descriptor for a molecule.
 
@@ -314,23 +416,22 @@ class PredictorDesc(DescriptorSet):
         Returns:
             a `list` of descriptor values
         """
-        mol = Chem.MolFromSmiles(mol) if isinstance(mol, str) else mol
-        return list(self._predictor.getScores([mol]))
+        return list(self._predictor.getScores(self.iterMols(mols, to_list=True)))
 
-    @property
+    @ property
     def is_fp(self):
         return self._is_fp
 
-    @property
+    @ property
     def settings(self):
         """Return args and kwargs used to initialize the descriptorset."""
-        return self._args, self._kwargs
+        return {"model_path": self.model_path, "metadata_path": self.metadata_path}
 
-    @property
+    @ property
     def descriptors(self):
         return self._descriptors
 
-    @descriptors.setter
+    @ descriptors.setter
     def descriptors(self, descriptors):
         self._descriptors = descriptors
 
@@ -354,8 +455,8 @@ class _DescriptorSetRetriever:
             raise Exception(f"{desc_type} is not a supported descriptor set type.")
         return method(*args, **kwargs)
 
-    def get_MorganFP(self, *args, **kwargs):
-        return MorganFP(*args, **kwargs)
+    def get_FingerprintSet(self, *args, **kwargs):
+        return FingerprintSet(*args, **kwargs)
 
     def get_DrugExPhyschem(self, *args, **kwargs):
         return DrugExPhyschem(*args, **kwargs)
@@ -368,6 +469,9 @@ class _DescriptorSetRetriever:
 
     def get_PredictorDesc(self, *args, **kwargs):
         return PredictorDesc(*args, **kwargs)
+
+    def get_TanimotoDistances(self, *args, **kwargs):
+        return TanimotoDistances(*args, **kwargs)
 
 
 def get_descriptor(desc_type: str, *args, **kwargs):