+++ CHANGELOG.md
@@ -25,7 +25,9 @@ From v1.0.0 to v1.1.0
 - Default `chunk_size` for `MoleculeTable.addDescriptors` was set to 50 so that smaller data sets can take advantage of more CPUs as well.
 - The `datasplit` interface was changed to mimic the `sklearn.model_selection.BaseCrossValidator` interface so all `sklearn` cross-validation methods can be used with QSPRPred out of the box to either generate train/test split or cross-validation splits (see the new features below)
 - Default `chunk_size` for `MoleculeTable` was set to 50 so that smaller data sets can take advantage of more CPUs as well.
-- The number of CPUs to use for parallel operations by `MoleculeTable` is now set in the `__init__` of the class
+- The number of CPUs to use for parallel operations  by `MoleculeTable` is now set in the `__init__` of the class
+- `DescriptorSets` are now initialized with the specific arguments instead of args and kwargs.
+- `MorganFP` was replaced by a more general class `FingerprintSet` which uses an object from the `Fingerprint` class as its fingerprint type
 - Passing invalid SMILES to the `getScores` function of  `Predictor` will now return a score of None. 
 
 ## New Features
@@ -37,5 +39,6 @@ From v1.0.0 to v1.1.0
 - The `datasplit` interface is now used to both generate train/test split and also the cross-validation splits
 - Train/test split of the data set is now saved in the matrix itself and is reloaded upon deserialization
 - `MoleculeTable` was updated with new features to generate scaffolds of molecules
+- `TanimotoDistances` was added as descriptortype.
 - Balanced class weighing was added as an option to the CLI
 - `PredictorDesc` was added as a new `DescriptorSet` type. It uses a QSPRpred model as descriptor.